Revert "Migrate op to new infra: paged_cache (#32874)"
This reverts commit 079b4860e78205e1e1ba451a08a33429737ac016.

This broke transformers APC tests: https://github.com/tenstorrent/tt-metal/actions/runs/19584682404/job/56094963419
