--- ttnn/cpp/ttnn/operations/experimental/paged_cache/device/fill_cache/paged_fill_cache_device_operation_types.hpp
@@ -1,29 +1,0 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "ttnn/tensor/tensor.hpp"
-#include "ttnn/operations/core/compute_kernel/compute_kernel_config.hpp"
-#include "ttnn/distributed/types.hpp"
-
-namespace ttnn::operations::experimental::paged_cache::fill {
-
-struct operation_attributes_t {
-    const uint32_t batch_idx_fallback;
-    const std::optional<std::set<ttnn::MeshCoordinate>> mesh_coords;
-};
-
-struct tensor_args_t {
-    Tensor cache_tensor;  // also output tensor
-    Tensor input_tensor;
-    Tensor page_table;
-    std::optional<Tensor> batch_idx_tensor_opt;
-};
-
-using tensor_return_value_t = Tensor;
-
-using spec_return_value_t = TensorSpec;
-
-}  // namespace ttnn::operations::experimental::paged_cache::fill
