+++ ttnn/cpp/ttnn/operations/experimental/paged_cache/device/paged_update_cache_program_factory.hpp
@@ -1,0 +1,21 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include <tt-metalium/constants.hpp>
+#include "ttnn/operations/cb_utils.hpp"
+#include "paged_cache_operation.hpp"
+#include <tt-metalium/work_split.hpp>
+
+namespace ttnn::operations::experimental::paged_cache::detail {
+
+tt::tt_metal::operation::ProgramWithCallbacks paged_update_cache_multi_core(
+    const Tensor& cache_tensor,
+    const Tensor& input_tensor,
+    std::optional<const Tensor> update_idxs_tensor,
+    std::optional<const Tensor> page_table,
+    const std::vector<uint32_t>& update_idxs,
+    uint32_t batch_offset,
+    ttnn::DeviceComputeKernelConfig compute_kernel_config,
+    bool share_cache);
+}  // namespace ttnn::operations::experimental::paged_cache::detail
