+++ ttnn/cpp/ttnn/operations/experimental/paged_cache/paged_cache.cpp
@@ -2,9 +2,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "device/update_cache/paged_update_cache_device_operation.hpp"
-#include "device/fused_update_cache/paged_fused_update_cache_device_operation.hpp"
-#include "device/fill_cache/paged_fill_cache_device_operation.hpp"
+#include "device/paged_cache_operation.hpp"
+#include "ttnn/run_operation.hpp"
 #include "ttnn/operations/core/compute_kernel/compute_kernel_config.hpp"
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/operations/core/core.hpp"
@@ -23,16 +22,23 @@ ttnn::Tensor PagedUpdateCacheOperation::invoke(
     const uint32_t batch_offset = 0,
     std::optional<const ttnn::DeviceComputeKernelConfig> compute_kernel_config = std::nullopt,
     const std::optional<const std::set<ttnn::MeshCoordinate>>& mesh_coords = std::nullopt) {
-    return ttnn::prim::paged_update_cache(
-        cache_tensor,
-        input_tensor,
-        update_idxs,
-        update_idxs_tensor,
-        share_cache,
-        page_table,
-        batch_offset,
-        compute_kernel_config,
-        mesh_coords);
+    auto kernel_config_val = init_device_compute_kernel_config(input_tensor.device()->arch(), compute_kernel_config);
+    const bool share_cache_arg = share_cache.has_value() ? share_cache.value() : false;
+    tt::tt_metal::operation::run(
+        PagedUpdateCacheDeviceOperation{
+            0,                               // .batch_idx_fallback (not used by UPDATE op type)
+            std::nullopt,                    // .batch_idx_tensor_opt (not used by UPDATE op type)
+            update_idxs,                     // .update_idxs
+            batch_offset,                    // .batch_offset
+            PagedUpdateCacheOpType::UPDATE,  // .op_type
+            kernel_config_val,               // .compute_kernel_config
+            share_cache_arg,                 // .share_cache
+            mesh_coords,                     // .mesh_coords
+        },
+        {cache_tensor, input_tensor},
+        {update_idxs_tensor, page_table});  // Optional inputs for UPDATE
+
+    return cache_tensor;  // Updated cache tensor in-place
 }
 
 std::tuple<ttnn::Tensor, ttnn::Tensor> PagedFusedUpdateCacheOperation::invoke(
@@ -47,18 +53,23 @@ std::tuple<ttnn::Tensor, ttnn::Tensor> PagedFusedUpdateCacheOperation::invoke(
     const uint32_t batch_offset = 0,
     std::optional<const ttnn::DeviceComputeKernelConfig> compute_kernel_config = std::nullopt,
     const std::optional<const std::set<ttnn::MeshCoordinate>>& mesh_coords = std::nullopt) {
-    return ttnn::prim::paged_fused_update_cache(
-        cache_tensor1,
-        input_tensor1,
-        cache_tensor2,
-        input_tensor2,
-        update_idxs,
-        update_idxs_tensor,
-        share_cache,
-        page_table,
-        batch_offset,
-        compute_kernel_config,
-        mesh_coords);
+    auto kernel_config_val = init_device_compute_kernel_config(input_tensor1.device()->arch(), compute_kernel_config);
+    const bool share_cache_arg = share_cache.has_value() ? share_cache.value() : false;
+    tt::tt_metal::operation::run(
+        PagedUpdateCacheDeviceOperation{
+            0,                                     // .batch_idx_fallback (not used by FUSED_UPDATE op type)
+            std::nullopt,                          // .batch_idx_tensor_opt (not used by FUSED_UPDATE op type)
+            update_idxs,                           // .update_idxs
+            batch_offset,                          // .batch_offset
+            PagedUpdateCacheOpType::FUSED_UPDATE,  // .op_type
+            kernel_config_val,                     // .compute_kernel_config
+            share_cache_arg,                       // .share_cache
+            mesh_coords,                           // .mesh_coords
+        },
+        {cache_tensor1, input_tensor1, cache_tensor2, input_tensor2},
+        {update_idxs_tensor, page_table});  // Optional inputs for FUSED_UPDATE
+
+    return {cache_tensor1, cache_tensor2};  // Updated cache tensors in-place
 }
 
 ttnn::Tensor PagedFillCacheOperation::invoke(
@@ -69,11 +80,24 @@ ttnn::Tensor PagedFillCacheOperation::invoke(
     const uint32_t batch_idx_fallback,
     std::optional<const ttnn::DeviceComputeKernelConfig> compute_kernel_config = std::nullopt,
     const std::optional<const std::set<ttnn::MeshCoordinate>>& mesh_coords = std::nullopt) {
-    // Note: compute_kernel_config is not used by fill_cache operation
-    (void)compute_kernel_config;
+    auto kernel_config_val = init_device_compute_kernel_config(input_tensor.device()->arch(), compute_kernel_config);
 
-    return ttnn::prim::paged_fill_cache(
-        cache_tensor, input_tensor, page_table, batch_idx_tensor, batch_idx_fallback, mesh_coords);
+    std::vector<std::optional<const Tensor>> optional_inputs_for_run;
+    tt::tt_metal::operation::run(
+        PagedUpdateCacheDeviceOperation{
+            batch_idx_fallback,            // .batch_idx_fallback (used by FILL if tensor not present)
+            batch_idx_tensor,              // .batch_idx_tensor_opt (used by FILL if present)
+            {},                            // .update_idxs (empty for fill)
+            0,                             // .batch_offset (0 for fill)
+            PagedUpdateCacheOpType::FILL,  // .op_type
+            kernel_config_val,             // .compute_kernel_config
+            false,        // .share_cache (false for fill, can be made a param if needed for future FILL variants)
+            mesh_coords,  // .mesh_coords (optional, can be used to restrict operation to specific mesh coordinates)
+        },
+        {cache_tensor, input_tensor, page_table},  // Mandatory inputs for FILL
+        {std::nullopt, std::nullopt});
+
+    return cache_tensor;  // Updated cache tensor in-place
 }
 
 }  // namespace operations::experimental::paged_cache
